new sendPacket pulling all common packet adjustments in one place
and with that, a temp. sendPreparedPacket for the activation call, since
that might not obey this pattern - TODO

With debugging on (not included in patches), it’s obvious that bytes:

0 = FF
1 = Data length
2 = Command
3 = Unknown, but replies include this number after AA
4 + onwards, Data
¤ = Checksum and end of data..

It does not appear that further data after the checksum matters, and
while can probably be unset should probably be 0 valued incase.