Move to using sendPacket for every command
This cleans up the code quite nicely, I like it.
I have not been able to replicate the packets commented, if the packet
length is indeed true then is this just random data from a mem dump and
therefore invalid?

Which also means the activation packet, if 17 bytes long, has more data
sent by us than needs be. Needs more investigation.

Unknown values in the colour commands also need to be explored, which
will be for another day.