Store device serial
It’s possible to query this over USB, but since we are given it in the
activation message, let us capture it there.

This instead could be fetched with:

if (descriptor.iSerialNumber) {
    char string[256];
    int ret = libusb_get_string_descriptor_ascii(tryDeviceHandler,
		descriptor.iSerialNumber, string, sizeof(string));
    int level = 2;
    if (ret > 0) {
	printf("Portal serial number: %s\n", string);
    }
}

however that provides the ugliness that is:
> 48E750913433 Performance Designed Products

Reasons for adding this is for potential multi-device uses, where you
wish to select which one to set LEDs on.