Support querying the active colour of a platform
This is a new command, 0x91, which takes a single parameter of the
platform like many of the other commands. However, since it does not
support 0 (meaning all platforms), 0 is used for platform 1, 1 for 2, 2
for 3.

It returns with 3 bytes of data for the request, RRGGBB.

(I realise since it is not mentioned in comments/docs, but yes 0 is a
valid platform number to set all at once)

This can be summed up by my comments on IRC (#linux on AfterNET)
2018-01-08 15:15:44 <@johndrinkwater> :D
2018-01-08 15:15:48 <@johndrinkwater> discovered a new command packet