Support platforms having blinking random colours
This is a new command, 0x94, which takes three parameters. One for the
platform (Supports 0, 1, 2, 3), one that appears to be duration of
pulse? and the last one for how many iterations before the command is
completed.

Having played a bit more, it seems we can set this command and
disconnect from the device (though obv. still leave it plugged in) and
it will iterate through the count you have requested. Though I have not
tested it, setting the largest values would make this command persist
for over an hour (16s * 256 iterations).

With this discovery, it seems we have more control over the lighting
behaviour of fadeColour/flashColour.. as the currently unknown values
appear to be timing/repetition values.

> squee
2018-01-08 16:21:38 <@johndrinkwater> found another.. flashes a random
colour with timing information you send, heh