Commit to naming the variables for this new command
Also renamed from flashRandomColours to fadeRandomColours, as it more
closely matches the behaviour of fadeColours.

It seems of the commands we have, we have an absolute sine wave, and a
pulse wave. It would not surprise me if further digging discovered more
patterns onboard the device.

Not entirely sure how to describe a variable that describe the time
period for a half period of a sine wave. It’s π but that aint a good
variable name, nor descriptive of the input the callee enters.

As described, putting 0x10 will result in an animation that lasts a
second before repeating.