Remove DeviceHandler.{cpp,h} as they are empty
Assuming these were meant to be agnostic classes around the portals,
if/when that is needed I am sure they can be returned.

Aiming to get ahold of the Lego base soon, and will see if that can join
these two to provide a decent lib for all toys to life LED+NFC bases.