Produced some more fuzzing output and retired old packet dumps
infinity-sample-packets is a collection of the packets in the previous
files, ordered by packet length, with repetitions removed and known
commands thinned out. This should help with combining the discovered
commands in infinity-discovered-commands-dump to produce actual working
interactions with the base.

There are some obvious commands that seem related to crypto/auth (0x8.)
cause they have rather asymmetric in/out buffers. There look to be more
animation commands in the higher 0x9., and there is a whole new set in
the 0xA. series that should be good to play with!

I did check up to 0xcf, but you never know there might be some at the
higher end.. for another poking session.