Support bulk setting of colours to all three platforms
This is a new command, 0x95, which takes (afaics) a byte-width bool to
confirm this change should apply, alongside 3 more bytes for the typical
RRGGBB setting.

Purpose of this is to update all platforms at the same time, while also
independently supplying unique colours to each. For the case where you
want to change just two platforms, this command is ideal.

Since some commands blend from initial colour to dst and back, I tested
sending a colour state (00FF00) yet also flagging it as no-set. The base
appears to ignore the colour in this case, not allowing us to influence
a fade from off (~000000) to the chosen colour.

For the purpose then, setting setPlayset/setPlayer? to 0, you are making
no change.

When you wish to set all platforms to the same single colour, prefer
setColour(0, …).