Support bulk setting sine waves on all three platforms
This is a new command, 0x96, which takes a byte-width bool to confirm
this change should apply, alongside the regular settings for sineWave(),
animationDuration, iterations, r, g, b, repeated 3 times, for all three
platforms.

It operates like sineWave(), but with the bonus of doing everything in
one packet rather than 3. Behaviour of the set variable is like the
commit description in ec43b9a841d577ab for setColours()