Support bulk setting pulse waves on all three platforms
This is a new command, 0x97, which takes a byte-width bool to confirm
this change should apply, alongside the regular settings for
pulseWave(), crest, trough, iterations, r, g, b, repeated 3 times, for
all three platforms.

It operates like pulseWave(), but with the bonus of doing everything in
one packet rather than 3. Behaviour of the set variable is like the
commit description in ec43b9a841d577ab for setColours()